+++ app/src/main/kotlin/com/erdenian/studentassistant/MainActivity.kt
@@ -8,7 +8,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import androidx.core.view.WindowCompat
 import com.erdenian.studentassistant.di.MainComponentHolder
 import com.erdenian.studentassistant.style.AppTheme
 
@@ -19,8 +18,6 @@ internal class MainActivity : AppCompatActivity() {
 
         var showSplashScreen by mutableStateOf(true)
         installSplashScreen().setKeepOnScreenCondition { showSplashScreen }
-
-        WindowCompat.setDecorFitsSystemWindows(window, false) // To make insets work
 
         setContent {
             LaunchedEffect(Unit) {