+++ core/style/src/main/kotlin/com/erdenian/studentassistant/style/Theme.kt
@@ -1,5 +1,7 @@
 package com.erdenian.studentassistant.style
 
+import android.app.Activity
+import android.content.ContextWrapper
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.icons.Icons
@@ -9,8 +11,15 @@ import androidx.compose.material3.dynamicLightColorScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.ReadOnlyComposable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.runtime.remember
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalInspectionMode
+import androidx.compose.ui.platform.LocalView
+import androidx.core.view.WindowCompat
 
 @Composable
 fun AppTheme(
@@ -26,6 +35,29 @@ fun AppTheme(
         (isDynamicColor && !isDarkTheme) -> dynamicLightColorScheme(LocalContext.current)
         (!isDynamicColor && isDarkTheme) -> DarkColorScheme
         else -> LightColorScheme
+    }
+
+    if (!LocalInspectionMode.current) {
+        val context = LocalContext.current
+        val window = remember(context) {
+            var currentContext = context
+            while (currentContext is ContextWrapper) {
+                if (currentContext is Activity) return@remember checkNotNull(currentContext.window)
+                currentContext = currentContext.baseContext
+            }
+            error("Activity not found")
+        }
+        val view = LocalView.current
+
+        SideEffect {
+            WindowCompat.setDecorFitsSystemWindows(window, false) // To make insets work
+
+            val insetsController = WindowCompat.getInsetsController(window, view)
+            window.statusBarColor = Color.Transparent.toArgb()
+            insetsController.isAppearanceLightStatusBars = !isDarkTheme
+            window.navigationBarColor = Color.Transparent.toArgb()
+            insetsController.isAppearanceLightNavigationBars = !isDarkTheme
+        }
     }
 
     CompositionLocalProvider(LocalDimensions provides dimensions) {